Fix type imports
